Move useUserApi up into New and Edit staff views
